user-service/requirements.txt
# 2024 amicroservice author.

# Core
grpcio==1.67.1
grpcio-tools==1.67.1
googleapis-common-protos==1.65.0
grpcio-status==1.67.1
bcrypt==4.2.0
PyJWT==2.9.0
protovalidate==0.5.0
asyncpg==0.30.0

# Test
isort==5.13.2
Faker==0.7.4
ipython==8.29.0
